+++ src/localops/tools/__init__.py
@@ -4,6 +4,7 @@ from .cpu import get_cpu_load
 from .disk import get_disk_usage
 from .errors import ToolCommandError
 from .memory import get_memory_usage
+from .services import get_service_status
 from .system import get_system_info
 
 __all__ = [
@@ -11,5 +12,6 @@ __all__ = [
     "get_cpu_load",
     "get_disk_usage",
     "get_memory_usage",
+    "get_service_status",
     "get_system_info",
 ]